+++ copy_js_bundle.sh
@@ -6,4 +6,3 @@ cd $BASEDIR
 rm -rf sdk/src/main/assets/
 mkdir -p sdk/src/main/assets/whiteboard
 cp -R ../whiteboard-bridge/build/* sdk/src/main/assets/whiteboard
-rabbit